use document statistics from pytorch-ie
