+++ configs/dataset/_collect_token_statistics.yaml
@@ -13,7 +13,7 @@ collect_token_numbers:
   format: markdown
   title: number of tokens
   metric:
-    _target_: src.metrics.statistics.DocumentTokenCounter
+    _target_: pytorch_ie.metrics.statistics.TokenCountCollector
     field: ${dataset.collect_token_statistics.text_field}
     tokenizer_name_or_path: ${dataset.collect_token_statistics.tokenizer_name_or_path}
     add_special_tokens: false
@@ -24,7 +24,7 @@ collect_token_numbers_hist:
   format: plot
   title: number of tokens distribution
   metric:
-    _target_: src.metrics.statistics.DocumentTokenCounter
+    _target_: pytorch_ie.metrics.statistics.TokenCountCollector
     field: ${dataset.collect_token_statistics.text_field}
     tokenizer_name_or_path: ${dataset.collect_token_statistics.tokenizer_name_or_path}
     add_special_tokens: false
